[CI] Fix Red bot version extract step
There is no `setup.py` so this commit adds a new field
`min_bot_version` in `info.json` and also adds a fallback version
of "3.4.0" in the workflow file iself.

See https://docs.discord.red/en/stable/guide_publish_cogs.html
for more details about `min_bot_version`. This field is supposed
to be defined for cogs only, not for cog repos, but at the very
least, it is a good central place to define the Red bot version
for the cog repo.